The schedule components organized into folders
